add global hooks for verification and refactor code
