+++ src/app.js
@@ -1,11 +1,12 @@
 import Fastify from "fastify";
 import ajvErrors from "ajv-errors";
-import authRoutes from "./modules/auth/infrastructure/routes/authRoutes.js";
+import authRoutes from "./modules/auth/infrastructure/authRoutes.js";
 import errorHandler from "./common/errorHandler.js";
 import fastifyJwt from "@fastify/jwt";
 import fastifySwagger from "@fastify/swagger";
 import fastifySwaggerUi from "@fastify/swagger-ui";
-import { jwtSecret } from "./common/config/config.js";
+import { jwtConfig } from "./common/config/config.js";
+import { registerHooks } from "./common/middlewares/hooks.js";
 import { swaggerOptions } from "./common/config/swagger.js";
 
 export async function createApp() {
@@ -24,7 +25,7 @@ export async function createApp() {
   fastify.setErrorHandler(errorHandler);
 
   // Register JWT plugin
-  fastify.register(fastifyJwt, { secret: jwtSecret });
+  fastify.register(fastifyJwt, { secret: jwtConfig.secret });
 
   // Register swagger
   fastify.register(fastifySwagger);
@@ -32,6 +33,9 @@ export async function createApp() {
 
   // Register api routes
   fastify.register(authRoutes, { prefix: "/api/auth" });
+
+  // Register global hooks
+  registerHooks(fastify);
 
   return fastify;
 }
