+++ src/common/config/config.js
@@ -1,4 +1,8 @@
 import dotenv from "dotenv";
+
 dotenv.config();
 
-export const jwtSecret = process.env.JWT_SECRET;
+export const jwtConfig = {
+  secret: process.env.JWT_SECRET,
+  expiresIn: "1h",
+};
