--- src/common/middlewares/authMiddleware.js
@@ -1,7 +1,0 @@
-export const authMiddleware = async (request, reply) => {
-  try {
-    await request.jwtVerify();
-  } catch (err) {
-    reply.code(401).send({ error: "Unauthorized" });
-  }
-};
