+++ src/common/utils/jwtUtils.js
@@ -1,10 +1,12 @@
+import { jwtConfig } from "../config/config.js";
+
 export default class JwtUtils {
   constructor(jwt) {
     this.jwt = jwt;
   }
 
   generateToken(payload) {
-    return this.jwt.sign(payload, { expiresIn: "1h" });
+    return this.jwt.sign(payload, { expiresIn: jwtConfig.expiresIn });
   }
 
   verifyToken(token) {
