+++ src/modules/auth/domain/auth.js
